+++ app/src/main/java/dev/rikoapp/cleanphonelauncher/LockDeviceAdminReceiver.kt
@@ -1,0 +1,5 @@
+package dev.rikoapp.cleanphonelauncher
+
+import android.app.admin.DeviceAdminReceiver
+
+class LockDeviceAdminReceiver : DeviceAdminReceiver()