+++ app/src/main/java/dev/rikoapp/cleanphonelauncher/presentation/home/HomeScreenAction.kt
@@ -17,4 +17,5 @@ sealed class HomeScreenAction {
     data class OnUninstallClick(val app: AppData) : HomeScreenAction()
     data class OnPhoneAppClick(val app: AppData) : HomeScreenAction()
     data class OnCameraAppClick(val app: AppData) : HomeScreenAction()
+    object OnDoubleTapHome : HomeScreenAction()
 }